updating ImGUI main
modified the result screen to display what is selected in the first windows sliders.
lowered the frame rounding
figured out how to make a background image. set century tower as background for now
added a quit and reset button on the top
modified the min and max values for sliders
changed the font to be more appealing